jump and collision are working
everything seems to be working. there are minor bugs like being able to
double jump a bit and some collisions seem to not work properly.

next step would be dynamic collision detection and to rework how
collisions are resolved. there is definitely a potential for a single
algorithm given the simplicity of the game.